SR: Portal Screen, Dimensional Generator (#292)
Portal Screen:
lists for challenges
more buttons

DG upgrades:
buttons have info on affordable status/cost
Changed behavior to match the rest of the game for SR: click is buy, 'tooltip' is info. (was click is info, buy is on a button several elements down. Works great visually, not so much with a screenreader)